requirements.txt
google-genai>=0.1.0
chromadb>=0.5.0
pandas>=2.0.0
numpy>=1.24.0
pymysql>=1.1.0
sqlparse>=0.5.0
python-dotenv>=1.0.0
tqdm>=4.66.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
loguru>=0.7.2


# Cloud Setup
google-cloud-storage>=2.14.0

# Multi-Agent & Testing
crewai>=0.1.0
crewai-tools>=0.1.0
langchain-google-genai>=0.0.1
pytest>=8.0.0
